+++ tests/TestCase/Command/Install/RunAllCommandTest.php
@@ -34,6 +34,7 @@ class RunAllCommandTest extends TestCase
     {
         $questions = (new RunAllCommand())->questions;
         $this->exec('me_cms.install -v', array_fill(0, count($questions), 'n'));
+        $this->assertExitWithSuccess();
 
         $expectedQuestions = array_column($questions, 'question');
         $outputQuestions = array_map(function (string $output): string {