+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.30 branch
 ### 2.30.5-RC2
 * fixed bootstrap, for loading other plugins;
+* fixed commands and their tests;
 * 'entities preferably use `hasValue()` instead of `has()`.
 
 ### 2.30.4-RC1